fix: Rename app to Lunark
