+++ app/src/androidTest/java/com/example/lunark/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.bookingapptim1;
+package com.example.lunark;
 
 import android.content.Context;
