+++ app/src/main/java/com/example/lunark/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.bookingapptim1;
+package com.example.lunark;
 
 import androidx.appcompat.app.AppCompatActivity;
